fix: first start
